FAI-628: Refactor use of TypedValue/TypedVariable
